backend/requirements.txt
# Backend Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# Google AI
google-genai==1.0.0

# Vector Search
faiss-cpu==1.9.0.post1
numpy==2.2.2

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
ruff==0.14.10